+++ parent/Parent_Report_Card.php
@@ -448,6 +448,6 @@ $studentSectionLabel = trim(($studentInfo['grade_level'] !== '' ? 'G' . $student
     </div>
 </div>
 <script src="<?php echo appAssetPath('vendor/bootstrap/bootstrap.bundle.min.js'); ?>"></script>
-<script src="../assets/js/main.js"></script>
+<script src="<?php echo appAssetPath('js/main.js'); ?>"></script>
 </body>
 </html>