+++ student/Student_Attendance.php
@@ -216,6 +216,6 @@ foreach ($classes as $class) {
     </div>
 </div>
 <script src="<?php echo appAssetPath('vendor/bootstrap/bootstrap.bundle.min.js'); ?>"></script>
-<script src="../assets/js/main.js"></script>
+<script src="<?php echo appAssetPath('js/main.js'); ?>"></script>
 </body>
 </html>