+++ student/Student_Classes.php
@@ -368,6 +368,6 @@ function formatFileSize($bytes) {
     </div>
 </div>
 <script src="<?php echo appAssetPath('vendor/bootstrap/bootstrap.bundle.min.js'); ?>"></script>
-<script src="../assets/js/main.js"></script>
+<script src="<?php echo appAssetPath('js/main.js'); ?>"></script>
 </body>
 </html>