+++ student/Student_Report_Card.php
@@ -370,6 +370,6 @@ $page_title = 'Report Card';
     </div>
 </div>
 <script src="<?php echo appAssetPath('vendor/bootstrap/bootstrap.bundle.min.js'); ?>"></script>
-<script src="../assets/js/main.js"></script>
+<script src="<?php echo appAssetPath('js/main.js'); ?>"></script>
 </body>
 </html>