+++ teacher/teacher_Advisory.php
@@ -463,7 +463,7 @@ $page_title = 'Report Card';
 </div>
 
 <script src="<?php echo appAssetPath('vendor/bootstrap/bootstrap.bundle.min.js'); ?>"></script>
-<script src="../assets/js/main.js"></script>
+<script src="<?php echo appAssetPath('js/main.js'); ?>"></script>
 <script>
 function submitReportCardToAdmin(academicYear){
     const fd=new FormData();