+++ teacher/teacher_Chat.php
@@ -264,7 +264,7 @@ $page_title = 'Messages';
     </div>
 
     <script src="<?php echo appAssetPath('vendor/bootstrap/bootstrap.bundle.min.js'); ?>"></script>
-    <script src="../assets/js/main.js"></script>
+    <script src="<?php echo appAssetPath('js/main.js'); ?>"></script>
     <script>
     const chatActionUrl = 'teacher_Chat_Action.php';
     const selectedParentId = <?php echo $selectedParentId; ?>;